+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The present file will list all changes made to the project; according to the
   The previous recipient option still exists if needed. This replacement will only happen once during the upgrade.
 
 #### Deprecated
+- Usage of "users_id_validate" parameter in "ajax/dropdownValidator.php" and "front/commonitilvalidation.form.php".
 - `front/change_problem.form.php` script usage.
 - `front/change_ticket.form.php` script usage.
 - `front/problem_ticket.form.php` script usage.